The localized textLayout.properties don't belong here.  Higher layers like Flex provide localized versions by using the TLF hook, GlobalSettings.resourceStringFunction to provide a function to that is called to get the localized string.  This hook is set in RichEditableText.initClass().
The GlobalSettings.resourceStringFunction function takes two parameters, a resource id and an optional array of parameters to substitute into the string.  The string is of form "Content {0} more content {1}". The parameters are read from the optional array and substituted for the bracketed substrings.  TLF provides a default implementation with default strings in English.  Clients may replace this function with their own implementation for localization.

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/tlf/branches/3.0@1412181 13f79535-47bb-0310-9956-ffa450edef68